rocblas cpu ref updated to ILP64 (#1761)
* rocblas cpu ref updated to ILP64

[ROCm/rocblas commit: da255afa5d81702877164ede786d4b08b424ea24]

+++ clients/include/host_vector.hpp
@@ -42,7 +42,7 @@ struct host_vector : std::vector<T, host_memory_allocator<T>>
     //! @brief Constructor.
     //! @param  inc Element index increment. If zero treated as one
     //!
-    host_vector(size_t n, ptrdiff_t inc = 1)
+    host_vector(size_t n, int64_t inc = 1)
         : std::vector<T, host_memory_allocator<T>>(n * std::abs(inc ? inc : 1))
         , m_n(n)
         , m_inc(inc ? inc : 1)
@@ -107,7 +107,7 @@ struct host_vector : std::vector<T, host_memory_allocator<T>>
     //!
     //! @brief Returns the increment of the vector.
     //!
-    ptrdiff_t inc() const
+    int64_t inc() const
     {
         return m_inc;
     }
@@ -137,6 +137,6 @@ struct host_vector : std::vector<T, host_memory_allocator<T>>
     }
 
 private:
-    size_t    m_n   = 0;
-    ptrdiff_t m_inc = 0;
+    size_t  m_n   = 0;
+    int64_t m_inc = 0;
 };
